Finalize Naive Cube Inference Algorithm
The cube inference algorithm is able to automatically infer the
structure of the cube given a data warehouse implemented in a RDBMS
using a snowflake schema. The algorithm has many restrictions and makes
a lot of assumptions about the structure of the data warehouse.
Made a change in gitignore as well.